+++ assets/patterns/help-users-when-we-time-them-out-of-a-service/timeoutDialog.js
@@ -1,215 +1,194 @@
 /* eslint-env jquery */
 
 require('jquery')
+var dialog = require('./dialog.js')
+var redirectHelper = require('./redirectHelper.js')
 
-Date.now = Date.now || function () { return +new Date() }
+var self = module.exports = {
+  timeoutDialog: function (options) {
 
-function secondsToTime (secs) {
-  var hours = Math.floor(secs / (60 * 60))
+    validateInput(options)
 
-  var divisorForMinutes = secs % (60 * 60)
-  var minutes = Math.floor(divisorForMinutes / 60)
-
-  var divisorForSeconds = divisorForMinutes % 60
-  var seconds = Math.ceil(divisorForSeconds)
-
-  var obj = {
-    'h': hours,
-    'm': minutes,
-    's': seconds
-  }
-  return obj
-}
-
-module.exports = function (options) {
-  var settings = {
-    timeout: 900,
-    countdown: 120,
-    time: 'minutes',
-    title: 'You’re about to be signed out',
-    message: 'For security reasons, you will be signed out of this service in',
-    keep_alive_url: '/keep-alive',
-    logout_url: '/sign-out',
-    restart_on_yes: true,
-    dialog_width: 340,
-    close_on_escape: true,
-    background_no_scroll: true,
-    keep_alive_button_text: 'Stay signed in',
-    sign_out_button_text: 'Sign out'
-  }
-
-  $.extend(settings, options)
-
-  var TimeoutDialog = {
-    init: function () {
-      this.setupDialogTimer()
-    },
-
-    setupDialogTimer: function () {
-      var self = this
-      self.dialogOpen = false
-      window.setTimeout(function () {
-        self.setupDialog()
-      }, ((settings.timeout) - (settings.countdown)) * 1000)
-    },
-
-    setupDialog: function () {
-      var self = this
-      window.dialogOpen = true
-      self.startTime = Math.round(Date.now() / 1000, 0)
-      self.currentMin = Math.ceil(settings.timeout / 60)
-      self.destroyDialog()
-      if (settings.background_no_scroll) {
-        $('html').addClass('noScroll')
+    var cleanupFunctions = []
+    var localisedDefaults = readCookie('PLAY_LANG') && readCookie('PLAY_LANG') === 'cy' && {
+      title: undefined,
+      message: 'Er eich diogelwch, byddwn yn eich allgofnodi cyn pen',
+      keepAliveButtonText: 'Parhau i fod wedi’ch mewngofnodi',
+      signOutButtonText: 'Allgofnodi',
+      properties: {
+        minutes: 'funud',
+        minute: 'funud',
+        seconds: 'eiliad',
+        second: 'eiliad'
       }
-      var time = secondsToTime(settings.countdown)
-      self.dialogOpen = true
-      if (time.m === 1) {
-        settings.time = ' minute'
+    } || {
+      title: undefined,
+      message: 'For your security, we will sign you out in',
+      keepAliveButtonText: 'Stay signed in',
+      signOutButtonText: 'Sign out',
+      properties: {
+        minutes: 'minutes',
+        minute: 'minute',
+        seconds: 'seconds',
+        second: 'second'
       }
-      $('<div id="timeout-dialog" class="timeout-dialog" role="dialog" aria-labelledby="timeout-message" tabindex=-1 aria-live="polite">' +
-        '<h1 class="heading-medium push--top">' + settings.title + '</h1>' +
-        '<p id="timeout-message" role="text">' + settings.message + ' <span id="timeout-countdown" class="countdown">' + time.m + ' ' + settings.time + '</span>' + '.</p>' +
-        '<button id="timeout-keep-signin-btn" class="button">' + settings.keep_alive_button_text + '</button>' +
-        '<a id="timeout-sign-out-btn" class="link">' + settings.sign_out_button_text + '</a>' +
-        '</div>' +
-        '<div id="timeout-overlay" class="timeout-overlay"></div>')
-      .appendTo('body')
-
-      // AL: disable the non-dialog page to prevent confusion for VoiceOver users
-      $('#skiplink-container, body>header, #global-cookie-message, body>main, body>footer').attr('aria-hidden', 'true')
-
-      var activeElement = document.activeElement
-      var modalFocus = document.getElementById('timeout-dialog')
-      modalFocus.focus()
-      self.addEvents()
-      self.startCountdown(settings.countdown)
-      self.escPress = function (event) {
-        if (event.keyCode === 27) {
-          // close the dialog
-          self.keepAlive()
-          activeElement.focus()
-        }
-      }
-
-      self.closeDialog = function () {
-        if (window.dialogOpen) {
-          self.keepAlive()
-          activeElement.focus()
-        }
-      }
-
-      // AL: prevent scrolling on touch, but allow pinch zoom
-      self.handleTouch = function (e) {
-        var touches = e.originalEvent.touches || e.originalEvent.changedTouches
-        if ($('#timeout-dialog').length) {
-          if (touches.length === 1) {
-            e.preventDefault()
-          }
-        }
-      }
-      // AL: add touchmove handler
-      $(document).on('touchmove', self.handleTouch)
-      $(document).on('keydown', self.escPress)
-      $('#timeout-keep-signin-btn').on('click', self.closeDialog)
-      $('#timeout-sign-out-btn').on('click', self.signOut)
-    },
-
-    destroyDialog: function () {
-      if ($('#timeout-dialog').length) {
-        window.dialogOpen = false
-        $('.timeout-overlay').remove()
-        $('#timeout-dialog').remove()
-        if (settings.background_no_scroll) {
-          $('html').removeClass('noScroll')
-        }
-      }
-      $('#skiplink-container, body>header, #global-cookie-message, body>main, body>footer').removeAttr('aria-hidden')
-    },
-
-    // AL: moved updater to own call to allow calling from other events
-    updateUI: function (counter) {
-      var self = this
-      if (counter < 60) {
-        $('.timeout-dialog').removeAttr('aria-live')
-        $('#timeout-countdown').html(counter + ' seconds')
-      } else {
-        var newCounter = Math.ceil(counter / 60)
-        var minutesMessage = ' minutes'
-        if (newCounter === 1) {
-          minutesMessage = ' minute'
-        }
-        if (newCounter < self.currentMin) {
-          self.currentMin = newCounter
-          $('#timeout-countdown').html(newCounter + minutesMessage)
-        }
-      }
-    },
-
-    addEvents: function () {
-      var self = this
-      // trap focus in modal (or browser chrome)
-      $('a, input, textarea, button, [tabindex]').not('[tabindex="-1"]').on('focus', function (event) {
-        var modalFocus = document.getElementById('timeout-dialog')
-        if (modalFocus && self.dialogOpen) {
-          if (!modalFocus.contains(event.target)) {
-            event.stopPropagation()
-            modalFocus.focus()
-          }
-        }
-      })
-
-      function handleFocus () {
-        if (self.dialogOpen) {
-          // clear the countdown
-          window.clearInterval(self.countdown)
-          // calculate remaining time
-          var expiredSeconds = (Math.round(Date.now() / 1000, 0)) - self.startTime
-          var currentCounter = settings.countdown - expiredSeconds
-          self.updateUI(currentCounter)
-          self.startCountdown(currentCounter)
-        }
-      }
-
-      // AL: handle browsers pausing timeouts/intervals by recalculating the remaining time on window focus
-      // need to widen this to cover the setTimeout which triggers the dialog for browsers which pause timers on blur
-      // hiding this from IE8 and it breaks the reset - to investigate further
-      if (navigator.userAgent.match(/MSIE 8/) == null) {
-        // $(window).on("blur", function(){
-        $(window).off('focus', handleFocus)
-        $(window).on('focus', handleFocus)
-        // });
-      }
-    },
-
-    startCountdown: function (counter) {
-      var self = this
-      self.countdown = window.setInterval(function () {
-        counter -= 1
-        self.updateUI(counter)
-        if (counter <= 0) {
-          self.signOut()
-        }
-      }, 1000)
-    },
-
-    keepAlive: function () {
-      var self = this
-      this.destroyDialog()
-      window.clearInterval(this.countdown)
-      $.get(settings.keep_alive_url, function () {
-        if (settings.restart_on_yes) {
-          self.setupDialogTimer()
-        } else {
-          self.signOut()
-        }
-      })
-    },
-
-    signOut: function () {
-      window.location = settings.logout_url
     }
-  }
 
-  TimeoutDialog.init()
+    var settings = mergeOptionsWithDefaults(options, localisedDefaults)
+
+    setupDialogTimer()
+
+    function validateInput(config) {
+      var requiredConfig = ['timeout', 'countdown', 'keepAliveUrl', 'signOutUrl']
+      var missingRequiredConfig = []
+
+      $.each(requiredConfig, function () {
+        if (!config.hasOwnProperty(this)) {
+          missingRequiredConfig.push(this)
+        }
+      })
+
+      if (missingRequiredConfig.length > 0) {
+        throw new Error('Missing config item(s): [' + missingRequiredConfig.join(', ') + ']')
+      }
+    }
+
+    function mergeOptionsWithDefaults(options, localisedDefaults) {
+      return $.extend({}, localisedDefaults, options)
+    }
+
+    function setupDialogTimer() {
+      settings.signout_time = getDateNow() + settings.timeout * 1000
+
+      var timeout = window.setTimeout(function () {
+        setupDialog()
+      }, ((settings.timeout) - (settings.countdown)) * 1000)
+
+      cleanupFunctions.push(function () {
+        window.clearTimeout(timeout)
+      })
+    }
+
+    function setupDialog() {
+      var $countdownElement = $('<span id="timeout-countdown" class="countdown">');
+      var $element = $('<div>')
+        .append(settings.title ? $('<h1 class="heading-medium push--top">').text(settings.title) : '')
+        .append($('<p id="timeout-message" role="text">').text(settings.message + ' ')
+          .append($countdownElement)
+          .append('.'))
+        .append($('<button id="timeout-keep-signin-btn" class="button">').text(settings.keepAliveButtonText))
+        .append($('<button id="timeout-sign-out-btn" class="button button--link">').text(settings.signOutButtonText))
+
+      $element.find('#timeout-keep-signin-btn').on('click', keepAliveAndClose)
+      $element.find('#timeout-sign-out-btn').on('click', signOut)
+
+      var dialogControl = dialog.displayDialog($element)
+
+      cleanupFunctions.push(function () {
+        dialogControl.closeDialog()
+      })
+
+      dialogControl.addCloseHandler(keepAliveAndClose)
+
+      dialogControl.setAriaLabelledBy('timeout-message')
+      if (getSecondsRemaining() > 60) {
+        dialogControl.setAriaLive('polite')
+      }
+
+      startCountdown($countdownElement, dialogControl)
+    }
+
+    function getSecondsRemaining() {
+      return Math.floor((settings.signout_time - getDateNow()) / 1000)
+    }
+
+    function startCountdown($countdownElement, dialogControl) {
+      function updateCountdown(counter, $countdownElement) {
+        var message
+        if (counter === 60) {
+          dialogControl.setAriaLive()
+        }
+        if (counter < 60) {
+          message = counter + ' ' + settings.properties[counter !== 1 ? 'seconds' : 'second']
+        } else {
+          var minutes = Math.ceil(counter / 60)
+          message = minutes + ' ' + settings.properties[minutes === 1 ? 'minute' : 'minutes']
+        }
+        $countdownElement.text(message)
+      }
+
+      function runUpdate() {
+        var counter = getSecondsRemaining()
+        updateCountdown(counter, $countdownElement)
+        if (counter <= 0) {
+          signOut()
+        }
+      }
+
+      var countdown = window.setInterval(runUpdate, 1000)
+      cleanupFunctions.push(function () {
+        window.clearInterval(countdown)
+      })
+      runUpdate()
+    }
+
+    function keepAliveAndClose() {
+      cleanup()
+      setupDialogTimer()
+      $.get(settings.keepAliveUrl, function () {
+      })
+    }
+
+    function getDateNow() {
+      return Date.now() || +new Date()
+    }
+
+    function signOut() {
+      redirectHelper.redirectToUrl(settings.signOutUrl)
+    }
+
+    function cleanup() {
+      while (cleanupFunctions.length > 0) {
+        var fn = cleanupFunctions.shift()
+        fn()
+      }
+    }
+
+    function readCookie(cookieName) { // From http://www.javascripter.net/faq/readingacookie.htm
+      var re = new RegExp('[; ]'+cookieName+'=([^\\s;]*)');
+      var sMatch = (' '+document.cookie).match(re);
+      if (cookieName && sMatch) return unescape(sMatch[1]);
+      return '';
+    }
+
+    return {cleanup: cleanup}
+  },
+  legacyWrapper: function (config) {
+    console.warn('$.timeout is now deprecated, please use window.GOVUK.timeoutDialog')
+    var updatedConfig = $.extend({}, config)
+
+    if (updatedConfig.hasOwnProperty('keep_alive_url')) {
+      updatedConfig.keepAliveUrl = updatedConfig.keep_alive_url
+      delete updatedConfig.keep_alive_url
+    }
+    if (updatedConfig.hasOwnProperty('keep_alive_button_text')) {
+      updatedConfig.keepAliveButtonText = updatedConfig.keep_alive_button_text
+      delete updatedConfig.keep_alive_button_text
+    }
+    if (updatedConfig.hasOwnProperty('logout_url')) {
+      updatedConfig.signOutUrl = updatedConfig.logout_url
+      delete updatedConfig.logout_url
+    }
+    if (updatedConfig.hasOwnProperty('sign_out_button_text')) {
+      updatedConfig.signOutButtonText = updatedConfig.sign_out_button_text
+      delete updatedConfig.sign_out_button_text
+    }
+
+    self.timeoutDialog($.extend({
+      timeout: 900,
+      countdown: 120,
+      keepAliveUrl: '/keep-alive',
+      signOutUrl: '/sign-out'
+    }, updatedConfig))
+  }
 }